+++ generators/app/templates/grunt/config/clean.js
@@ -1,0 +1,30 @@
+/**
+ * Generated header by C0ZEN for generator-altran-angular project
+ * Generated file clean on WebStorm
+ *
+ * Created by: Geoffrey "C0ZEN" Testelin
+ * Date: 19/11/2017
+ * Time: 21:17
+ * Version: 1.0.0
+ */
+/* eslint key-spacing:"off" */
+module.exports = {
+	server   : '.tmp',
+	release  : {
+		files: [
+			{
+				dot: true,
+				src: [
+					'.tmp',
+					'<%%= currentTarget %%>/**/*'
+				]
+			}
+		]
+	},
+	languages: [
+		'.tmp/languages'
+	],
+	devIndex : [
+		'<%%= paths.app %%>/index.html'
+	]
+};