+++ generators/app/templates/grunt/config/htmlmin.js
@@ -1,0 +1,9 @@
+/**
+ * Generated header by C0ZEN for generator-altran-angular project
+ * Generated file htmlmin on WebStorm
+ *
+ * Created by: Geoffrey "C0ZEN" Testelin
+ * Date: 19/11/2017
+ * Time: 22:35
+ * Version: 1.0.0
+ */