+++ generators/app/templates/grunt/config/jsMin.js
@@ -1,0 +1,13 @@
+/**
+ * Generated header by C0ZEN for generator-altran-angular project
+ * Generated file jsMin on WebStorm
+ *
+ * Created by: Geoffrey "C0ZEN" Testelin
+ * Date: 19/11/2017
+ * Time: 22:37
+ * Version: 1.0.0
+ */
+module.exports = {
+	dev    : null,
+	release: null
+};