+++ generators/app/templates/grunt/config/notify.js
@@ -1,0 +1,42 @@
+/**
+ * Generated header by C0ZEN for generator-altran-angular project
+ * Generated file notify on WebStorm
+ *
+ * Created by: Geoffrey "C0ZEN" Testelin
+ * Date: 19/11/2017
+ * Time: 22:36
+ * Version: 1.0.0
+ */
+/* eslint key-spacing:"off" */
+module.exports = {
+	test      : {
+		options: {
+			title  : '\'TEST\'',
+			message: 'This is just a test !'
+		}
+	},
+	serve     : {
+		options: {
+			title  : '\'SERVE\'',
+			message: 'Task complete ! Enjoy the app !'
+		}
+	},
+	release   : {
+		options: {
+			title  : '\'RELEASE\'',
+			message: 'Task complete !'
+		}
+	},
+	buildReady: {
+		options: {
+			title  : '\'BUILD READY\'',
+			message: 'The last build was successful ! You can now refresh your browser.'
+		}
+	},
+	less      : {
+		options: {
+			title  : '\'LESS UPDATED\'',
+			message: 'LESS files are now updated'
+		}
+	}
+};