+++ generators/app/templates/grunt/config/useminPrepare.js
@@ -1,0 +1,26 @@
+/**
+ * Generated header by C0ZEN for generator-altran-angular project
+ * Generated file useminPrepare on WebStorm
+ *
+ * Created by: Geoffrey "C0ZEN" Testelin
+ * Date: 19/11/2017
+ * Time: 21:34
+ * Version: 1.0.0
+ */
+/* eslint key-spacing:"off" */
+module.exports = {
+	html   : '<%%= currentTarget %%>/index.html',
+	options: {
+		dest: '<%%= currentTarget %%>',
+		flow: {
+			steps: {
+				js : [
+					'concat'
+				],
+				css: [
+					'concat'
+				]
+			}
+		}
+	}
+};