+++ generators/app/templates/grunt/config/wiredep.js
@@ -1,0 +1,18 @@
+/**
+ * Generated header by C0ZEN for generator-altran-angular project
+ * Generated file wiredep on WebStorm
+ *
+ * Created by: Geoffrey "C0ZEN" Testelin
+ * Date: 19/11/2017
+ * Time: 21:20
+ * Version: 1.0.0
+ */
+/* eslint key-spacing:"off" */
+module.exports = {
+	app: {
+		src       : [
+			'<%%= paths.app %%>/config/tpls/index.tpl.html'
+		],
+		ignorePath: /\.\.\//
+	}
+};